requirements.txt
flask==3.0.0
langchain==1.0.5
langchain-community==0.4.1
langchain-google-genai==3.0.3
langchain-chroma==1.0.0
chromadb==1.3.4
pillow==11.3.0
pypdf2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
python-dotenv==1.0.0
